Fix test errors after rebase
